% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca}
\alias{cca}
\title{Calculates the total Corrected Covered Area (CCA) Index}
\usage{
cca(cm)
}
\arguments{
\item{cm}{A dataframe for the citation matrix}
}
\value{
res
}
\description{
This package calculates the Corrected Covered Area (CCA) index. The measure for assessing the overall degree of overlap in an OOSR. It is taking as input the citation matrix.
}
\examples{
DATASET <- readxl::read_excel(system.file('extdata','cca.xlsx', package = 'ccaR'))

tb <- cca(DATASET)



}
