% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt}
\alias{dt}
\title{The dataset from the publication}
\format{
A data frame with 14 rows and 6 variables:
\describe{
  \item{Chasan2014}{The publication of Chasan2014}
  \item{Gilinsky2015}{The publication of Gilinsky2015}
  \item{Guo2016}{The publication of Guo2016}
  \item{Middleton2014}{The publication of Middleton2014}
  \item{Morton2014}{The publication of Morton2014}
  \item{Peacock2014}{The publication of Peacock2014}
}
}
\source{
\url{https://hsda.med.auth.gr/}
}
\usage{
data(dt)
}
\description{
The dataset contaoineing the publications
}
\keyword{datasets}
