% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca}
\alias{cca}
\title{Calculates the overall Corrected Covered Area (CCA) Index}
\usage{
cca(cm)
}
\arguments{
\item{cm}{Defines the data frame containing 1s, 0s, and NAs (in case of structural missingness).}
}
\value{
res
}
\description{
It calculates the overall CCA index for the entire citation matrix. It is taking as input the citation matrix.
}
\examples{
DATASET <- readxl::read_excel(system.file('extdata','cca.xlsx', package = 'ccaR'))

tb <- cca(DATASET)



}
