% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccaR.R
\name{cca}
\alias{cca}
\title{Calculates the corrected covered area (CCA) index}
\usage{
cca(cm)
}
\arguments{
\item{cm}{A dataframe for the citation matrix}
}
\value{
res
}
\description{
This package calculates the corrected covered area (CCA) index. The measure for assessing the overall degree of overlap in an OoSR. It is taking as input the citation matrix.
}
\examples{
DATASET<-read.csv(system.file('extdata','miyazaki2017CCA.csv', package = 'ccaR') ,sep = ";")
ccaM<-cca(DATASET)

}
