% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_heatmap.R
\name{cca_heatmap}
\alias{cca_heatmap}
\title{Plots the cca heatmap}
\usage{
cca_heatmap(cm, fontsize = 5, chroma = "#527e11")
}
\arguments{
\item{cm}{A dataframe for the citation matrix}

\item{fontsize}{A number which controls the aesthetic of font size of the numbers in the tiles of the heatmap}

\item{chroma}{The color of the heatmap}
}
\value{
cca_heatmap
}
\description{
This function plots the cca heatmap. The tiles within the triangular matrix contain color-coded data that demonstrate the degree of overlap between pairs of reviews using the corrected covered area (CCA) measure. It is taking as input the citation matrix, the font size of the text in the tiles and the color used in the heatmap.
}
\examples{
DATASET<-readxl::read_excel(system.file('extdata','cca.xlsx', package = 'ccaR'))


cca_heatmap(DATASET, 3) +
    ggplot2::theme(
        plot.caption = ggplot2::element_text(size = 16, margin=ggplot2::margin(30,0,0,0)),
        legend.title = ggplot2::element_text(size = 16, face = "bold", vjust=4),
        legend.text = ggplot2::element_text(size = 16),
        legend.key.size = ggplot2::unit(1.0, "cm"),
        legend.title.align = 0.5,
        legend.text.align = 0.5,
        axis.text.x=ggplot2::element_text(size = 16),
        axis.text.y=ggplot2::element_text(size = 16),
        axis.title=ggplot2::element_blank(),
        axis.ticks=ggplot2::element_blank(),
        axis.line=ggplot2::element_blank(),
        panel.border=ggplot2::element_blank(),
        panel.grid.major.x=ggplot2::element_line(colour = "grey80", linetype = "dashed"))
}
