% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_table.R
\name{cca_table}
\alias{cca_table}
\title{Calculates the Corrected Covered Area (CCA) Index}
\usage{
cca_table(cm)
}
\arguments{
\item{cm}{A dataframe for the citation matrix}
}
\value{
res
}
\description{
This package calculates the Corrected Covered Area (CCA) index. The measure for assessing the overall degree of overlap in an OOSR. It is taking as input the citation matrix.
}
\examples{
DATASET <- readxl::read_excel(system.file('extdata','cca.xlsx', package = 'ccaR'))

tb <- cca_table(DATASET)



}
